chore(release): 对齐生产基线 guide-20260803124914 + 全链路 v3.3
- 同步 guide 基线: agent_responses/user_capability(113能力)/page_guide_service 等
- 保留全链路改动(fullchain/knowledge/attack_surface/adversarial),与其共存不冲突
- 同步 capability 计数测试(109→113)

+++ backend/app/models/project_source_archive.py
@@ -23,7 +23,7 @@ class ProjectSourceArchive(Base, IdMixin, TimestampMixin):
     original_filename = Column(String(255), nullable=False)
     media_type = Column(String(120), nullable=False, default="application/zip")
     archive_sha256 = Column(String(64), nullable=False)
-    compressed_size = Column(BigInteger, nullable=False)
+    compressed_size = Column(Integer, nullable=False)
     expanded_size = Column(BigInteger, nullable=False)
     file_count = Column(Integer, nullable=False)
     max_member_size = Column(BigInteger, nullable=False)
